@@ -8,10 +8,10 @@ import (
 func serve() {
 	http.HandleFunc("/user", func(w http.ResponseWriter, r *http.Request) {
 		r.ParseForm()
-		username := r.Form.Get("username")
+		username := r.Form.Get("username") // $ Source[go/reflected-xss]
 		if !isValidUsername(username) {
 			// BAD: a request parameter is incorporated without validation into the response
-			fmt.Fprintf(w, "%q is an unknown user", username)
+			fmt.Fprintf(w, "%q is an unknown user", username) // $ Alert[go/reflected-xss]
 		} else {
 			// TODO: Handle successful login
 		}
